fix(ios): resolve Kotlin/Native compile errors (category imports, atomics, expect/actual)

+++ composeApp/src/iosMain/kotlin/com/nuvio/app/features/player/StillWatchingPlatform.ios.kt
@@ -6,5 +6,5 @@ import platform.Foundation.NSCalendarUnitHour
 
 internal actual fun currentHour(): Int {
     val calendar = NSCalendar.currentCalendar
-    return calendar.component(NSCalendarUnitHour, fromDate = NSDate())
+    return calendar.component(NSCalendarUnitHour, fromDate = NSDate()).toInt()
 }
